commented out doublon tests + moved upper/lower divertor tests to test_BallReactor
